[modified] form of variable | [add] section: use of AI

+++ main.c
@@ -6,7 +6,7 @@
 /*   By: sonoe <sonoe@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/02/28 05:19:21 by soyamagu          #+#    #+#             */
-/*   Updated: 2026/04/19 20:34:51 by sonoe            ###   ########.fr       */
+/*   Updated: 2026/04/19 21:15:57 by sonoe            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ int	main(int argc, char **argv)
 	stack_b = NULL;
 	if (argc < 2)
 		return (escape_error(2));
-	stack_a = parse_args(argv + 1, argc - 1, &stack_a);
+	stack_a = parse_args((argv + 1), (size_t)(argc - 1), &stack_a);
 	if (!stack_a)
 		return (escape_error(2));
 	compress_data(&stack_a);
